refactor(navigation): make collections the profile tab
Move the profile collection entry point into the main tab bar as My, relocate settings behind the timeline toolbar, and keep the timeline profile menu focused on web-aligned profile destinations.
